Add files via upload
I added the functionality to send direct messages, the functionality to request to see the user's unread direct messages, and the functionality to request to see all of the user's direct messages. These functionalities were stated as such in the assignment prompt on Canvas. However, the second and third functionalities list (to read the direct messages) does not work yet.